Basic fade works on individual channels with hard-coded remote codes

+++ types.h
@@ -1,7 +1,7 @@
 typedef struct {
   float valNow;
   float val0;
-  float t0;
+  unsigned long t0;
   float pin;
 } Colour;
